backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
redis[hiredis]==5.2.1
h3==4.1.1
websockets==14.1
orjson==3.10.12
pydantic-settings==2.7.1
shapely==2.0.6
rasterio>=1.3.10
numpy>=1.26.0
scipy>=1.12.0
aiohttp>=3.9.0
aiofiles>=23.2.0
numba>=0.59.0
imageio[pyav]>=2.34.0
fpdf2>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
